+++ src/openfacefx/__init__.py
@@ -59,6 +59,7 @@ from .export_unreal_notifies import write_unreal_notifies, notifies_to_dict
 from .export_live2d import write_live2d_motion, lipsync_param_ids
 from .export_godot import write_godot_anim
 from .export_gltf import write_gltf, build_gltf
+from .export_vrma import write_vrma, build_vrma, VRM_EXPRESSION_PRESETS
 from .export_vmd import write_vmd, vmd_bytes, DEFAULT_MORPH_MAP
 from .export_livelink import (write_livelink_csv, livelink_csv_string,
                               LIVELINK_COLUMNS)
@@ -141,6 +142,7 @@ __all__ = [
     "write_unreal_notifies", "notifies_to_dict",
     "write_live2d_motion", "lipsync_param_ids", "write_godot_anim",
     "write_gltf", "build_gltf",
+    "write_vrma", "build_vrma", "VRM_EXPRESSION_PRESETS",
     "write_vmd", "vmd_bytes", "DEFAULT_MORPH_MAP",
     "write_livelink_csv", "livelink_csv_string", "LIVELINK_COLUMNS",
     "write_lip", "lip_bytes", "skyrim_mapping", "SKYRIM_SLOT_MAP",